requirement.txt
fastapi
uvicorn
openai
requests
sentence-transformers
faiss-cpu
spacy
PyPDF2
streamlit